[KLC-1007] Review G115 (Possible Overflows) & Enable SonarQube Report (#807)
* review possible int convert overflows and ignore remaining for gosec report (G115)

* scenarios timeout with env

* update actions google-auth v2

* enable sonarqube scan

* update gosec fixed release

* golangci-lint

* fix operations validate builtin

* generate hooks name and importer

* remove store2elastic

* Fix tests

* golangci && Rearrange git workflow

* refactor proposal create

* refactor checkBlockValidity

* refactor loadBlocks

* refactor addSmartContract

* refactor branchNode delete

* refactor CreateTransaction

* refactor vm test

* refactor node CreateTransaction

* remove unused k8s

* sonar fix

* update git workflow cache

* ErrInvalidReceiverAddress dup code

+++ health/healthService.go
@@ -131,7 +131,8 @@ func (h *healthService) monitorMemory() {
 
 	log.Trace("healthService.monitorMemory()", "heapInUse", tools.ConvertBytes(stats.HeapInuse))
 
-	if int(stats.HeapInuse) > h.config.MemoryUsageToCreateProfiles {
+	// #nosec G115
+	if stats.HeapInuse > uint64(h.config.MemoryUsageToCreateProfiles) {
 		recordObj := newMemoryUsageRecord(stats, h.clock.now(), h.folder)
 		h.records.addRecord(recordObj)
 	}
